+++ functions/model/tasks/createTask.js
@@ -1,5 +1,5 @@
-/* eslint-disable no-unused-vars */
 /* eslint-disable linebreak-style */
+/* eslint-disable no-unused-vars */
 /* eslint-disable require-jsdoc */
 /* eslint-disable object-curly-spacing */
 // /* eslint-disable no-undef */
