+++ src/app/app.module.ts
@@ -44,8 +44,6 @@ import { RangeCardComponent } from "./body/my-dashboard/performance-chart/range-
 import { CreateNewOrganizationComponent } from './body/create-new-organization/create-new-organization.component'
 import { ErrorComponent } from './body/error/error.component';
 import { Patch1Component } from './patches/patch1/patch1.component';
-import { OrganizationDetailsComponent } from './body/create-new-organization/organization-details/organization-details.component';
-import { TeamDetailsComponent } from './body/create-new-organization/team-details/team-details.component';
 import { FocusNavComponent } from './body/board/focus-nav/focus-nav.component';
 import { UserVerificationComponent } from './user-verification/user-verification.component';
 import { ProfileComponent } from './body/profile/profile.component';
@@ -62,7 +60,7 @@ import { UpdateFieldComponent } from './patches/patch3/update-field/update-field
 import { ProfileIconsComponent } from './icons/profile-icons/profile-icons.component';
 import { ProgressGraphComponent } from './body/common-tools/progress-graph/progress-graph.component';
 import { SuggestionBucketComponent } from './body/common-tools/suggestion-bucket/suggestion-bucket.component';
-import { AddMemberComponent } from './body/create-new-organization/team-details/add-member/add-member.component';
+import { AddMemberComponent } from './body/create-new-team/add-member/add-member.component';
 import { PerformanceColumnChartComponent } from './body/board/performance-column-chart/performance-column-chart.component';
 import { ColumnChartComponent } from './body/common-tools/column-chart/column-chart.component';
 import { SelectDropdownComponent } from './body/board/performance-column-chart/select-dropdown/select-dropdown.component';
@@ -75,6 +73,7 @@ import { DifficultyIconsComponent } from './icons/difficulty-icons/difficulty-ic
 import { ListPatchesComponent } from './list-patches/list-patches.component';
 import { PatchCardsComponent } from './list-patches/patch-cards/patch-cards.component';
 import { EditPatchComponent } from './list-patches/edit-patch/edit-patch.component';
+import { CreateNewTeamComponent } from './body/create-new-team/create-new-team.component';
 
 
 @NgModule({
@@ -109,8 +108,6 @@ import { EditPatchComponent } from './list-patches/edit-patch/edit-patch.compone
     ErrorComponent,
     FilterTaskComponent,
     Patch1Component,
-    OrganizationDetailsComponent,
-    TeamDetailsComponent,
     FocusNavComponent,
     UserVerificationComponent,
     ProfileComponent,
@@ -139,7 +136,8 @@ import { EditPatchComponent } from './list-patches/edit-patch/edit-patch.compone
     DifficultyIconsComponent,
     ListPatchesComponent,
     PatchCardsComponent,
-    EditPatchComponent
+    EditPatchComponent,
+    CreateNewTeamComponent
   ],
   imports: [
     BrowserModule,
