--- src/app/body/create-new-organization/organization-details/organization-details.component.css
@@ -1,68 +1,0 @@
-.input {
-    border-radius: 10px;
-    padding: 10px;
-    border: var(--primary-border);
-    outline: none;
-    width: 100%;
-}
-
-#createNewTaskContainer {
-    border: var(--primary-border);
-    background-color: var(--secondary-bg);
-    color: var(--text);
-}
-
-#uploadFile {
-    color: var(--text);
-    border: var(--primary-border);
-    border-radius: 10px;
-}
-
-input[type="file"] {
-    display: none;
-}
-
-#important {
-    color: red;
-}
-
-#button {
-    background-color: var(--secondary-color);
-    border: var(--primary-border);
-    color: var(--button-text);
-}
-
-#button:focus,
-#button:active,
-#submitButton:focus,
-#submitButton:active {
-    outline: 0 none;
-    box-shadow: none;
-}
-
-#submitButton {
-    background-color: transparent;
-    color: var(--text);
-    border: none;
-}
-
-#submitButton:hover {
-    cursor: pointer;
-    color: var(--primary-color);
-}
-
-#submitButton:disabled {
-    color: gray;
-}
-
-#navbarMenu {
-    background-color: var(--primary-color);
-}
-
-select:focus,
-input:focus,
-textarea:focus {
-    box-shadow: none;
-    outline: 0 none;
-    border-color: var(--primary-color);
-}
