--- src/app/body/create-new-organization/organization-details/organization-details.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { OrganizationDetailsComponent } from './organization-details.component';
-
-describe('OrganizationDetailsComponent', () => {
-  let component: OrganizationDetailsComponent;
-  let fixture: ComponentFixture<OrganizationDetailsComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ OrganizationDetailsComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(OrganizationDetailsComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
