+++ src/app/body/create-new-team/create-new-team.component.css
@@ -27,6 +27,7 @@
 select:focus,
 input:focus,
 textarea:focus {
-    box-shadow: 0px 0px 5px 0px #dc3226;
+    box-shadow: none;
     outline: 0 none;
+    border-color: var(--primary-color);
 }
