+++ src/app/body/login/login.component.html
@@ -5,11 +5,6 @@
         </div>
     </div>
     <div class="row align-items-center justify-content-center">
-        <div class="col-12 col-sm-6 col-md-6 text-center p-5">
-            <button type="submit" class="btn" id="button" (click)="onSignUpWithOrg()">
-                Create an Organization
-            </button>
-        </div>
         <div class="col-12 col-sm-6 col-md-6 p-2" id="containerBorder">
             <div class="row" id="emailForm">
                 <div class="col">
